remove: rename includes in dir greeting
