+++ front/src/app/pages/greeting/greeting.scss
